move files, fix text colors
